clean up persoenlicheDaten interface, move related data to shared location
