+++ app/flows/beratungshilfeFormular/abgabe/guards.ts
@@ -1,9 +1,9 @@
+import { persoenlicheDatenDone } from "~/flows/shared/persoenlicheDaten/doneFunctions";
 import { type BeratungshilfeFormularContext } from "..";
 import type { Guards } from "../../guards.server";
 import { anwaltlicheVertretungDone } from "../anwaltlicheVertretung/guards";
 import { beratungshilfeFinanzielleAngabeDone } from "../finanzielleAngaben/beratungshilfeFinanzielleAngabeDone";
 import { grundvoraussetzungDone } from "../grundvoraussetzung/context";
-import { beratungshilfePersoenlicheDatenDone } from "../persoenlicheDaten/context";
 import { rechtsproblemDone } from "../rechtsproblem/context";
 
 export const beratungshilfeAbgabeGuards = {
@@ -12,7 +12,7 @@ export const beratungshilfeAbgabeGuards = {
     anwaltlicheVertretungDone({ context }) &&
     rechtsproblemDone({ context }) &&
     beratungshilfeFinanzielleAngabeDone({ context }) &&
-    beratungshilfePersoenlicheDatenDone({ context }),
+    persoenlicheDatenDone({ context }),
   abgabeOnline: ({ context }) => context.abgabeArt == "online",
   abgabeAusdrucken: ({ context }) => context.abgabeArt == "ausdrucken",
 } satisfies Guards<BeratungshilfeFormularContext>;
