+++ app/flows/beratungshilfeFormular/context.ts
@@ -1,15 +1,15 @@
 import { beratungshilfeAnwaltlicheVertretung } from "./anwaltlicheVertretung/context";
 import { beratungshilfeFinanzielleAngaben } from "./finanzielleAngaben/context";
 import { beratungshilfeGrundvoraussetzungen } from "./grundvoraussetzung/context";
-import { beratungshilfePersoenlicheDaten } from "./persoenlicheDaten/context";
 import { beratungshilfeRechtsproblem } from "./rechtsproblem/context";
 import { abgabeContext } from "../shared/abgabe/context";
+import { persoenlicheDaten } from "~/flows/shared/persoenlicheDaten/context";
 
 export const beratungshilfeFormularContext = {
   ...beratungshilfeAnwaltlicheVertretung,
   ...beratungshilfeGrundvoraussetzungen,
   ...beratungshilfeRechtsproblem,
   ...beratungshilfeFinanzielleAngaben,
-  ...beratungshilfePersoenlicheDaten,
+  ...persoenlicheDaten,
   ...abgabeContext,
 } as const;
