+++ app/flows/fluggastrechteFormular/persoenlicheDaten/context.ts
@@ -1,12 +1,8 @@
+import _ from "lodash";
 import { z } from "zod";
-import {
-  adresseSchema,
-  namePrivatPerson,
-} from "~/flows/shared/persoenlicheDaten/context";
+import { persoenlicheDaten as sharedPersoenlicheDaten } from "~/flows/shared/persoenlicheDaten/context";
 import { pageDataSchema } from "~/services/flow/pageDataSchema";
 import { checkedOptional } from "~/services/validation/checkedCheckbox";
-import { optionalOrSchema } from "~/services/validation/optionalOrSchema";
-import { phoneNumberSchema } from "~/services/validation/phoneNumber";
 import { postcodeSchema } from "~/services/validation/postcode";
 import { stringRequiredSchema } from "~/services/validation/stringRequired";
 import { YesNoAnswer } from "~/services/validation/YesNoAnswer";
@@ -20,19 +16,17 @@ const fluggastrechtePersoenlichVertretungDaten = {
   beschreibenVertretung: stringRequiredSchema,
 };
 
+const persoenlicheDaten = _.omit(sharedPersoenlicheDaten, ["geburtsdatum"]);
+
 export const fluggastrechtePersoenlichDaten = {
-  ...namePrivatPerson,
-  ...adresseSchema,
-  telefonnummer: optionalOrSchema(phoneNumberSchema),
+  ...persoenlicheDaten,
   unter18JahreAlt: checkedOptional,
   ...fluggastrechtePersoenlichVertretungDaten,
   isWeiterePersonen: YesNoAnswer,
   weiterePersonen: z.array(
     z
       .object({
-        ...namePrivatPerson,
-        ...adresseSchema,
-        telefonnummer: optionalOrSchema(phoneNumberSchema),
+        ...persoenlicheDaten,
         unter18JahreAlt: checkedOptional,
         ...fluggastrechtePersoenlichVertretungDaten,
       })
